refactor: strings (#16)
