+++ utils/strings.ts
@@ -1,9 +1,9 @@
-import es from "~/i18n/es.js";
-import en from "~/i18n/en.js";
+import es from "~/strings/es";
+import en from "~/strings/en";
 
 const locales: Record<string, Record<string, string>> = { es, en };
 
-class I18n {
+class Strings {
   code: string;
   constructor (code: string) {
     this.code = String(code).toLowerCase();
@@ -18,8 +18,8 @@ class I18n {
   }
 }
 
-export const i18n = new I18n("es");
+export const strings = new Strings("es");
 
 export const t = (key: string) => {
-  return i18n.get(key);
+  return strings.get(key);
 };
